Refactor radio schedule utilities for reusing in cypress tests
