+++ src/app/containers/RadioSchedule/utilities/processRadioSchedule.js
@@ -1,8 +1,7 @@
-import findLastIndex from 'ramda/src/findLastIndex';
-import propSatisfies from 'ramda/src/propSatisfies';
 import path from 'ramda/src/path';
 import nodeLogger from '#lib/logger.node';
 import { RADIO_SCHEDULE_DATA_INCOMPLETE_ERROR } from '#lib/logger.const';
+import { getLastProgramIndex, isScheduleDataComplete } from '.';
 
 const logger = nodeLogger(__filename);
 
@@ -29,23 +28,6 @@ const logProgramError = ({ error, service }) => {
     error,
     service,
   });
-};
-
-const getLastProgramIndex = ({ schedules, currentTime }) => {
-  return findLastIndex(
-    propSatisfies(time => time < currentTime, 'publishedTimeStart'),
-  )(schedules);
-};
-
-export const isScheduleDataComplete = ({
-  schedules,
-  currentTime = Date.now(),
-}) => {
-  // finding latest program, that may or may not still be live. this is because there isn't
-  // always a live program, in which case we show the most recently played program on demand.
-  const latestProgramIndex = getLastProgramIndex({ schedules, currentTime });
-
-  return schedules[latestProgramIndex - 2] && schedules[latestProgramIndex + 1];
 };
 
 export default (data, service, currentTime) => {
